Organize library paths
