+++ lib/package.json
@@ -1,6 +1,6 @@
 {
   "name": "dooboo-ui",
-  "version": "0.0.13",
+  "version": "0.0.15",
   "main": "index.js",
   "types": "index.d.ts",
   "author": "dooboolab",
