+++ main/__tests__/Button.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import Button from '../Button';
+import { Button } from '../../main';
 import { Text } from 'react-native';
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
