+++ main/index.ts
@@ -1,8 +1,11 @@
-export * from './Accordion';
-export * from './Button';
-export * from './ButtonGroup';
-export * from './EditText';
-export * from './LoadingIndicator';
-export * from './Slider';
-export * from './SwitchToggle';
-export * from './Select';
+export { Accordion } from './Accordion';
+export { Button } from './Button';
+export { ButtonGroup } from './ButtonGroup';
+export { EditText, EditTextInputType } from './EditText';
+export { LoadingIndicator } from './LoadingIndicator';
+export { Slider } from './Slider';
+export { RadioButton } from './RadioButton';
+export { Rating } from './Rating';
+export { SearchInput } from './SearchInput';
+export { SwitchToggle } from './SwitchToggle';
+export { Select, SelectItem } from './Select';
