+++ stories/dooboo-ui/ButtonGroup.stories.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement, useState } from 'react';
 import { Text, View } from 'react-native';
 
-import ButtonGroup from '../../main/ButtonGroup';
+import { ButtonGroup } from '../../main';
 import { ContainerDeco } from '../../storybook/decorators';
 import { storiesOf } from '@storybook/react-native';
 import styled from 'styled-components/native';
