+++ stories/dooboo-ui/LoadingIndicator.stories.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from 'react';
 
 import { ContainerDeco } from '../../storybook/decorators';
-import LoadingIndicator from '../../main/LoadingIndicator';
+import { LoadingIndicator } from '../../main';
 import { storiesOf } from '@storybook/react-native';
 import styled from 'styled-components/native';
