+++ stories/dooboo-ui/RadioButton.stories.tsx
@@ -3,7 +3,7 @@ import { ScrollView, View } from 'react-native';
 import { boolean, color, number, select, text } from '@storybook/addon-knobs';
 
 import { ContainerDeco } from '../../storybook/decorators';
-import RadioButton from '../../main/RadioButton';
+import { RadioButton } from '../../main';
 import { storiesOf } from '@storybook/react-native';
 import styled from 'styled-components/native';
