+++ stories/dooboo-ui/Rating.stories.tsx
@@ -3,7 +3,7 @@ import React, { ReactElement } from 'react';
 import { boolean, number } from '@storybook/addon-knobs';
 
 import { ContainerDeco } from '../../storybook/decorators';
-import Rating from '../../main/Rating';
+import { Rating } from '../../main';
 import { storiesOf } from '@storybook/react-native';
 import styled from 'styled-components/native';
