+++ stories/dooboo-ui/SearchInput.stories.tsx
@@ -1,8 +1,9 @@
 import React, { useState } from 'react';
 import { SafeAreaView, ScrollView } from 'react-native';
 import { number, text } from '@storybook/addon-knobs';
+
 import { IC_MAGNIFIER } from '../Icon';
-import SearchInput from '../../main/SearchInput';
+import { SearchInput } from '../../main';
 import { storiesOf } from '@storybook/react-native';
 import styled from 'styled-components/native';
