+++ stories/dooboo-ui/Slider.stories.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement, useState } from 'react';
 
 import { ContainerDeco } from '../../storybook/decorators';
-import Slider from '../../main/Slider';
+import { Slider } from '../../main';
 import { Text } from 'react-native';
 import { storiesOf } from '@storybook/react-native';
 import styled from 'styled-components/native';
